refactor(reservationTimeRepository): 이용 가능 시간 조회 책임 이동
