+++ src/main/java/roomescape/theme/repository/ThemeRepository.java
@@ -1,6 +1,5 @@
 package roomescape.theme.repository;
 
-import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 import roomescape.theme.domain.Theme;
@@ -16,6 +15,4 @@ public interface ThemeRepository {
     List<Theme> findAll();
 
     boolean existsByName(String name);
-
-    List<AvailableTimeQueryResult> findAvailableTimes(Long themeId, LocalDate date);
 }
