+++ src/test/java/roomescape/time/repository/JdbcReservationTimeRepositoryTest.java
@@ -2,6 +2,7 @@ package roomescape.time.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
@@ -14,7 +15,10 @@ import roomescape.time.domain.ReservationTime;
 @JdbcTest
 class JdbcReservationTimeRepositoryTest {
 
-    private final ReservationTimeRepository reservationTimeRepository;
+    @Autowired
+    JdbcTemplate jdbcTemplate;
+
+    ReservationTimeRepository reservationTimeRepository;
 
     @Autowired
     public JdbcReservationTimeRepositoryTest(JdbcTemplate jdbcTemplate) {
@@ -94,5 +98,33 @@ class JdbcReservationTimeRepositoryTest {
         // then
         assertThat(exists).isTrue();
         assertThat(notExists).isFalse();
+    }
+
+    @Test
+    @DisplayName("예약되지 않은 시간만 조회된다")
+    void findAvailableTimes() {
+        // given
+        Long themeId = 1L;
+        LocalDate date = LocalDate.of(2025, 1, 1);
+
+        jdbcTemplate.update("insert into reservation_time(id, start_at) values (1, '10:00:00')");
+        jdbcTemplate.update("insert into reservation_time(id, start_at) values (2, '11:00:00')");
+        jdbcTemplate.update("insert into reservation_time(id, start_at) values (3, '12:00:00')");
+
+        jdbcTemplate.update("insert into theme(id, name, description, thumbnail_url) values (1, '테마', '설명', 'url')");
+
+        jdbcTemplate.update("""
+            insert into reservation(name, reservation_date, time_id, theme_id)
+            values ('user', ?, 2, ?)
+        """, date, themeId);
+
+        // when
+        List<AvailableTimeQueryResult> result = reservationTimeRepository.findAvailableTimes(themeId, date);
+
+        // then
+        assertThat(result).hasSize(2);
+        assertThat(result)
+                .extracting(AvailableTimeQueryResult::startAt)
+                .containsExactlyInAnyOrder(LocalTime.of(10, 0), LocalTime.of(12, 0));
     }
 }
